% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{axivity_close}
\alias{axivity_close}
\title{Close a serial connection to an Axivity device}
\usage{
axivity_close(handle)
}
\arguments{
\item{handle}{A connection handle returned by \code{\link[=axivity_open]{axivity_open()}}.}
}
\description{
Close a serial connection to an Axivity device
}
