% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{axivity_download_cancel}
\alias{axivity_download_cancel}
\title{Cancel an in-progress download from an Axivity device}
\usage{
axivity_download_cancel(device_id)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}
}
\description{
Cancel an in-progress download from an Axivity device
}
