% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{axivity_download_status}
\alias{axivity_download_status}
\alias{axivity_download_wait}
\title{Check or wait for an Axivity device's download progress}
\usage{
axivity_download_status(device_id)

axivity_download_wait(device_id)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}
}
\value{
A list with \code{status} (\code{"none"}, \code{"error"}, \code{"progress"},
\code{"complete"}, or \code{"cancelled"}) and \code{value} (percentage complete if
\code{status} is \code{"progress"}, a diagnostic code if \code{"error"}).
}
\description{
Check or wait for an Axivity device's download progress
}
