% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{axivity_get_accelerometer}
\alias{axivity_get_accelerometer}
\title{Read an Axivity device's current accelerometer values}
\usage{
axivity_get_accelerometer(device_id)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}
}
\value{
A named numeric vector \code{c(x, y, z)} in units of \emph{g}
(raw values are in 1/256 \emph{g}, converted here).
}
\description{
Read an Axivity device's current accelerometer values
}
