% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{axivity_get_battery}
\alias{axivity_get_battery}
\title{Query an Axivity device's battery level and health}
\usage{
axivity_get_battery(device_id)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}
}
\value{
A list with \code{level_pct} (0-99\% = charging, 100\% = full) and
\code{recharge_cycles} (lower is better).
}
\description{
Query an Axivity device's battery level and health
}
