% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{axivity_get_ecc}
\alias{axivity_get_ecc}
\alias{axivity_set_ecc}
\title{Get or set an Axivity device's error-correcting code (ECC) flag}
\usage{
axivity_get_ecc(device_id)

axivity_set_ecc(device_id, enabled)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}

\item{enabled}{Logical. Enable or disable ECC.}
}
\value{
\code{axivity_get_ecc()} returns a logical.
}
\description{
Get or set an Axivity device's error-correcting code (ECC) flag
}
