% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{axivity_set_led}
\alias{axivity_set_led}
\title{Set an Axivity device's LED colour}
\usage{
axivity_set_led(device_id, colour)
}
\arguments{
\item{device_id}{Integer device ID, as returned by \code{\link[=axivity_discover]{axivity_discover()}}.}

\item{colour}{One of \code{"auto"} (default device-controlled behaviour),
\code{"off"}, \code{"blue"}, \code{"green"}, \code{"cyan"}, \code{"red"}, \code{"magenta"},
\code{"yellow"}, \code{"white"}.}
}
\description{
Set an Axivity device's LED colour
}
